Breaking up the integration test suite for the table
